remove regular CallableResolver, rename AdvancedCallableResolver and fix tests
