+++ Slim/MiddlewareDispatcher.php
@@ -177,48 +177,55 @@ class MiddlewareDispatcher implements RequestHandlerInterface
 
             public function handle(ServerRequestInterface $request): ResponseInterface
             {
-                $callable = null;
                 if ($this->callableResolver instanceof AdvancedCallableResolverInterface) {
                     $callable = $this->callableResolver->resolveMiddleware($this->middleware);
-                } else {
-                    if ($this->callableResolver instanceof CallableResolverInterface) {
-                        try {
-                            $callable = $this->callableResolver->resolve($this->middleware);
-                        } catch (RuntimeException $ignore) {
-                        }
+                    return $callable($request, $this->next);
+                }
+
+                $callable = null;
+
+                if ($this->callableResolver instanceof CallableResolverInterface) {
+                    try {
+                        $callable = $this->callableResolver->resolve($this->middleware);
+                    } catch (RuntimeException $e) {
+                        // Do Nothing
+                    }
+                }
+
+                if (!$callable) {
+                    $resolved = $this->middleware;
+                    $instance = null;
+                    $method = null;
+
+                    // Check for Slim callable as `class:method`
+                    if (preg_match(CallableResolver::$callablePattern, $resolved, $matches)) {
+                        $resolved = $matches[1];
+                        $method = $matches[2];
                     }
 
-                    if (!is_callable($callable)) {
-                        $resolved = $this->middleware;
-                        $instance = null;
-                        $method = null;
-
-                        // Check for Slim callable as `class:method`
-                        if (preg_match(CallableResolver::$callablePattern, $resolved, $matches)) {
-                            $resolved = $matches[1];
-                            $method = $matches[2];
-                        }
-
-                        if ($this->container && $this->container->has($resolved)) {
-                            $instance = $this->container->get($resolved);
-                            if ($instance instanceof MiddlewareInterface) {
-                                return $instance->process($request, $this->next);
-                            }
-                        } elseif (!function_exists($resolved)) {
-                            if (!class_exists($resolved)) {
-                                throw new RuntimeException(sprintf('Middleware %s does not exist', $resolved));
-                            }
-                            $instance = new $resolved($this->container);
-                        }
-
-                        if ($instance && $instance instanceof MiddlewareInterface) {
+                    if ($this->container && $this->container->has($resolved)) {
+                        $instance = $this->container->get($resolved);
+                        if ($instance instanceof MiddlewareInterface) {
                             return $instance->process($request, $this->next);
                         }
-
-                        $callable = $instance ?? $resolved;
-                        if ($instance && $method) {
-                            $callable = [$instance, $method];
+                    } elseif (!function_exists($resolved)) {
+                        if (!class_exists($resolved)) {
+                            throw new RuntimeException(sprintf('Middleware %s does not exist', $resolved));
                         }
+                        $instance = new $resolved($this->container);
+                    }
+
+                    if ($instance && $instance instanceof MiddlewareInterface) {
+                        return $instance->process($request, $this->next);
+                    }
+
+                    $callable = $instance ?? $resolved;
+                    if ($instance && $method) {
+                        $callable = [$instance, $method];
+                    }
+
+                    if ($this->container && $callable instanceof Closure) {
+                        $callable = $callable->bindTo($this->container);
                     }
                 }
 
@@ -227,10 +234,6 @@ class MiddlewareDispatcher implements RequestHandlerInterface
                         'Middleware %s is not resolvable',
                         $this->middleware
                     ));
-                }
-
-                if ($this->container && $callable instanceof Closure) {
-                    $callable = $callable->bindTo($this->container);
                 }
 
                 return $callable($request, $this->next);
