+++ tests/TestCase.php
@@ -18,9 +18,7 @@ use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\StreamFactoryInterface;
 use Psr\Http\Message\StreamInterface;
 use Psr\Http\Server\RequestHandlerInterface;
-use Slim\AdvancedCallableResolver;
 use Slim\CallableResolver;
-use Slim\Interfaces\AdvancedCallableResolverInterface;
 use Slim\Interfaces\CallableResolverInterface;
 use Slim\MiddlewareDispatcher;
 use Slim\Tests\Providers\PSR7ObjectProvider;
@@ -65,35 +63,20 @@ abstract class TestCase extends PhpUnitTestCase
     }
 
     /**
-     * Get a default advanced callable resolver with or without a container
-     *
-     * @param ContainerInterface|null $container
-     *
-     * @return AdvancedCallableResolverInterface
-     */
-    protected function getAdvancedCallableResolver(
-        ?ContainerInterface $container = null
-    ): AdvancedCallableResolverInterface {
-        return new AdvancedCallableResolver($container);
-    }
-
-    /**
      * @param RequestHandlerInterface $requestHandler
      * @param ContainerInterface|null $container
-     * @param bool                    $useAdvancedCallableResolver
+     * @param CallableResolverInterface|null $callableResolver
      *
      * @return MiddlewareDispatcher
      */
     protected function createMiddlewareDispatcher(
         RequestHandlerInterface $requestHandler,
         ?ContainerInterface $container = null,
-        bool $useAdvancedCallableResolver = true
+        ?CallableResolverInterface $callableResolver = null
     ): MiddlewareDispatcher {
         return new MiddlewareDispatcher(
             $requestHandler,
-            $useAdvancedCallableResolver ?
-                $this->getAdvancedCallableResolver($container) :
-                $this->getCallableResolver($container),
+            $callableResolver ?? $this->getCallableResolver($container),
             $container
         );
     }
